Nest embarcadero's Image/Oci satellites and enigmatic's Cipher.Session
`ImageConfig`, `ImageHandle` and `OciError` become `Image.Config`, `Image.Handle`
and `Oci.Error`; `CipherSession` becomes `Cipher.Session`, under a new companion
for `trait Cipher`. Four names leave the umbrella's exported surface.

`BlockCipherMode` and `BlockCipherPadding` stay whole, for the reason profanity's
`TerminalEvent` stays whole: each donor file holds a taxonomy — `Cbc`, `Ecb`,
`Ctr`, `Cfb`, `Ofb` and `Pkcs7`, `Iso10126`, `NoPadding` — whose members are
exported single-word names in their own right. The second pass's rule that a
taxonomy moves into the namespace whole holds only when its members are not
independently exported; otherwise nesting the container silently renames all of
them.
