+++ lib/embarcadero/src/oci/embarcadero.Image.scala
@@ -32,6 +32,10 @@
                                                                                                   */
 package embarcadero
 
+import aperture.*
+import fulminate.*
+import jacinta.*
+import pneumatic.*
 import proscenium.compat.*
 
 import anticipation.*
@@ -44,17 +48,19 @@ import hieroglyph.*, charEncoders.utf8Encoder
 import hypotenuse.*
 import prepositional.*
 import rudiments.map
+import scala.caps
 import serpentine.*
 import spectacular.*
 import turbulence.*
 import vacuous.*
 import wisteria.*
+import zephyrine.*
 
 object Image:
   // Anchored here so `data.open[Image]()` resolves with no import. Opening a filesystem
   // *path* as an image (`path.open[Image]`) lives in the JVM-only source set.
   given dataOpenable
-  :   ( ociTactic: Tactic[OciError], tarTactic: Tactic[Tar.Error],
+  :   ( ociTactic: Tactic[Oci.Error], tarTactic: Tactic[Tar.Error],
         streamTactic: Tactic[StreamError] )
   =>  ( ImageDataOpenable^{ociTactic, tarTactic, streamTactic} ) =
     ImageDataOpenable()
@@ -70,7 +76,7 @@ object Image:
   :   Image =
 
     val imageConfig =
-      ImageConfig
+      Image.Config
         ( architecture = architecture,
           os           = os,
           rootfs       = RootFs(t"layers", layers.map(_.diffId)),
@@ -141,6 +147,153 @@ object Image:
     Image(layers, config, configBytes, configDescriptor, manifest, manifestBytes,
         manifestDescriptor, index, indexBytes)
 
+  // ImageConfig → Image.Config
+  // An OCI image config document (media type `MediaType.config`). It records the
+  // target platform, the rootfs `diff_id` chain, and the optional runtime
+  // configuration and build history.
+  case class Config
+    ( architecture: Text,
+      os:           Text,
+      rootfs:       RootFs,
+      config:       Optional[ContainerConfig] = Unset,
+      history:      Optional[List[History]]   = Unset,
+      created:      Optional[Text]            = Unset,
+      author:       Optional[Text]            = Unset )
+  derives CanEqual
+
+  // ImageHandle → Image.Handle
+  // The scoped capability provided by opening an oci-archive as `Image`:
+  // `path.open[Image]()`. The entries are materialized from bitumen's sequential reader,
+  // so blobs are found by name over the entry list; their memoized bodies are retained
+  // for the scope's duration, the flat cost of TAR's sequentiality. Archives written by
+  // `Image.archive` place `index.json` before the blobs, so the metadata path is cheap.
+  class Handle private[embarcadero] (entries: List[Tar.Entry])
+  extends caps.ExclusiveCapability:
+
+    // The body of the blob addressed by a canonical `sha256:<hex>` digest: its stored
+    // (for layers: compressed) chunks — undecoded and unverified.
+    private def body(digest: Text)(using Tactic[Oci.Error]): Tar.Body =
+      if !digest.s.startsWith("sha256:")
+      then abort(Oci.Error(Oci.Error.Reason.UnsupportedDigest(digest.cut(t":").stdlib.head)))
+
+      val name = t"blobs/sha256/${digest.s.stripPrefix("sha256:").tt}"
+
+      entries.stdlib.collectFirst { case file: Tar.Entry.File if file.entryName == name => file.data }
+      . getOrElse(abort(Oci.Error(Oci.Error.Reason.MissingBlob(digest))))
+
+    // The blob addressed by a canonical `sha256:<hex>` digest, as a stream of its stored
+    // (for layers: compressed) chunks — undecoded and unverified. (An explicit `Tactic`
+    // rather than `raises` sugar: a fresh capability in a context-function result cannot
+    // flow to a forwarding caller.)
+    def blob(digest: Text)(using Tactic[Oci.Error]): (Stream[Data] over Credit)^ =
+      body(digest).stream
+
+    // The decoded top-level index, after validating the `oci-layout` marker.
+    def index(using Tactic[Oci.Error]): Index =
+      val layoutBytes = document(t"oci-layout", Oci.Error.Reason.MissingLayout)
+
+      val layout = decode(t"oci-layout"):
+        import strategies.throwUnsafely
+        layoutBytes.read[Json].as[OciLayout]
+
+      if !layout.imageLayoutVersion.s.startsWith("1.")
+      then abort(Oci.Error(Oci.Error.Reason.UnsupportedLayout(layout.imageLayoutVersion)))
+
+      val indexBytes = document(t"index.json", Oci.Error.Reason.MissingIndex)
+
+      decode(t"index.json"):
+        import strategies.throwUnsafely
+        indexBytes.read[Json].as[Index]
+
+    // The index's first manifest, or the one a descriptor selects; the manifest blob is
+    // digest-verified against its descriptor before decoding.
+    def manifest(using Tactic[Oci.Error]): Oci.Manifest =
+      val descriptor =
+        index.manifests.stdlib.headOption.getOrElse(abort(Oci.Error(Oci.Error.Reason.NoManifest)))
+
+      manifest(descriptor)
+
+    def manifest(descriptor: Descriptor)(using Tactic[Oci.Error]): Oci.Manifest =
+      val bytes = verified(descriptor)
+
+      decode(descriptor.digest):
+        import strategies.throwUnsafely
+        bytes.read[Json].as[Oci.Manifest]
+
+    // The decoded image config for a manifest (by default, the first).
+    def imageConfig(using Tactic[Oci.Error]): Image.Config = imageConfig(manifest)
+
+    def imageConfig(manifest: Oci.Manifest)(using Tactic[Oci.Error]): Image.Config =
+      val bytes = verified(manifest.config)
+
+      decode(manifest.config.digest):
+        import strategies.throwUnsafely
+        bytes.read[Json].as[Image.Config]
+
+    // The decoded Wasm artifact config for a manifest (by default, the first).
+    def wasmConfig(using Tactic[Oci.Error]): WasmConfig = wasmConfig(manifest)
+
+    def wasmConfig(manifest: Oci.Manifest)(using Tactic[Oci.Error]): WasmConfig =
+      val bytes = verified(manifest.config)
+
+      decode(manifest.config.digest):
+        import strategies.throwUnsafely
+        bytes.read[Json].as[WasmConfig]
+
+    // The decoded config blob for a manifest (by default, the first), in whichever form the
+    // config descriptor's media type says it takes. This is the entry point for a reader
+    // that does not already know which kind of artifact it has opened — a runtime deciding
+    // whether to unpack a rootfs or instantiate a component.
+    def config(using Tactic[Oci.Error]): Oci.Config = config(manifest)
+
+    def config(manifest: Oci.Manifest)(using Tactic[Oci.Error]): Oci.Config =
+      if manifest.config.mediaType == media"application/vnd.wasm.config.v0+json"
+      then wasmConfig(manifest) else imageConfig(manifest)
+
+    // A layer's stored blob, verbatim: for OCI layers, the gzip-compressed tar.
+    def compressed(descriptor: Descriptor)(using Tactic[Oci.Error])
+    :   (Stream[Data] over Credit)^ =
+
+      blob(descriptor.digest)
+
+    // A layer's content as the uncompressed tar byte stream, decompressing according to
+    // the descriptor's media type; unrecognised types stream verbatim.
+    def layer(descriptor: Descriptor)(using Tactic[Oci.Error]): (Stream[Data] over Credit)^ =
+      if descriptor.mediaType.suffixes.stdlib.contains(Media.Suffix.Gzip)
+      then compressed(descriptor).decompress[Gzip]
+      else compressed(descriptor)
+
+    // A blob gathered eagerly and checked against its descriptor's digest and size — the
+    // opt-in verified path, since checking a stream would force draining it.
+    def verified(descriptor: Descriptor)(using Tactic[Oci.Error]): Data =
+      val bytes = body(descriptor.digest).memoize
+      val digest = sha256(bytes)
+
+      if digest != descriptor.digest
+      then abort(Oci.Error(Oci.Error.Reason.DigestMismatch(descriptor.digest, digest)))
+
+      if bytes.length.toLong != descriptor.size
+      then abort:
+        Oci.Error:
+          Oci.Error.Reason.InvalidBlob
+            ( descriptor.digest, t"its size is ${bytes.length}, not ${descriptor.size}" )
+
+      bytes
+
+    // The gathered bytes of a named top-level document (`oci-layout` or `index.json`).
+    private def document(name: Text, reason: Oci.Error.Reason)(using Tactic[Oci.Error]): Data =
+      entries.stdlib.collectFirst { case file: Tar.Entry.File if file.entryName == name => file.data }
+      . getOrElse(abort(Oci.Error(reason)))
+      . memoize
+
+    // Runs a JSON decode, translating any failure — parse, JSON or media-type errors,
+    // thrown under the call site's `throwUnsafely` — to an `InvalidBlob` on the given
+    // label. The decoder is derived under a throwing strategy because the derivation's
+    // codec thunks cannot capture a scoped tactic capability.
+    private def decode[doc](label: Text)(body: => doc)(using Tactic[Oci.Error]): doc =
+      try body catch case error: Error =>
+        abort(Oci.Error(Oci.Error.Reason.InvalidBlob(label, error.message.text)))
+
 case class Image
   ( layers:             List[Layer],
     config:             Oci.Config,
@@ -155,7 +308,7 @@ case class Image
   // The config blob narrowed to whichever form it takes; each is `Unset` for an image of
   // the other kind, so a caller that knows which it built can read the typed document
   // without a match.
-  def imageConfig: Optional[ImageConfig] = config.only { case config: ImageConfig => config }
+  def imageConfig: Optional[Image.Config] = config.only { case config: Image.Config => config }
   def wasmConfig: Optional[WasmConfig] = config.only { case config: WasmConfig => config }
 
   // Every blob in the image, as `(digest, bytes)` pairs: the config, each layer,
