+++ lib/embarcadero/src/oci/soundness_embarcadero_oci.scala
@@ -32,5 +32,5 @@
                                                                                                   */
 package soundness
 
-export embarcadero.{ContainerConfig, Descriptor, History, Image, ImageConfig, ImageDataOpenable,
-    ImageHandle, Index, Layer, Oci, OciError, RootFs, WasmComponent, WasmConfig}
+export embarcadero.{ContainerConfig, Descriptor, History, Image, ImageDataOpenable, Index, Layer,
+    Oci, RootFs, WasmComponent, WasmConfig}
